[chore] Move list offsets to admin
