+++ crates/fluss/src/client/table/scanner.rs
@@ -22,14 +22,12 @@ use crate::metadata::{TableBucket, TableInfo, TablePath};
 use crate::proto::{FetchLogRequest, PbFetchLogReqForBucket, PbFetchLogReqForTable};
 use crate::record::{LogRecordsBatchs, ReadContext, ScanRecord, ScanRecords, to_arrow_schema};
 use crate::rpc::RpcClient;
-use crate::rpc::message::{ListOffsetsRequest, OffsetSpec};
 use crate::util::FairBucketStatusMap;
 use parking_lot::RwLock;
 use std::collections::HashMap;
 use std::slice::from_ref;
 use std::sync::Arc;
 use std::time::Duration;
-use tokio::task::JoinHandle;
 
 const LOG_FETCH_MAX_BYTES: i32 = 16 * 1024 * 1024;
 #[allow(dead_code)]
@@ -67,7 +65,6 @@ pub struct LogScanner {
     metadata: Arc<Metadata>,
     log_scanner_status: Arc<LogScannerStatus>,
     log_fetcher: LogFetcher,
-    conns: Arc<RpcClient>,
 }
 
 impl LogScanner {
@@ -88,7 +85,6 @@ impl LogScanner {
                 metadata.clone(),
                 log_scanner_status.clone(),
             ),
-            conns: connections.clone(),
         }
     }
 
@@ -104,103 +100,6 @@ impl LogScanner {
         self.log_scanner_status
             .assign_scan_bucket(table_bucket, offset);
         Ok(())
-    }
-
-    pub async fn list_offsets_latest(&self, buckets: Vec<i32>) -> Result<HashMap<i32, i64>> {
-        // TODO: support partition_id
-        let partition_id = None;
-        let offset_spec = OffsetSpec::Latest;
-
-        self.metadata
-            .check_and_update_table_metadata(from_ref(&self.table_path))
-            .await?;
-
-        let cluster = self.metadata.get_cluster();
-        let table_id = cluster.get_table(&self.table_path).table_id;
-
-        // Prepare requests
-        let requests_by_server = self.prepare_list_offsets_requests(
-            table_id,
-            partition_id,
-            buckets.clone(),
-            offset_spec,
-        )?;
-
-        // Send Requests
-        let response_futures = self.send_list_offsets_request(requests_by_server).await?;
-
-        let mut results = HashMap::new();
-
-        for response_future in response_futures {
-            let offsets = response_future.await.map_err(
-                // todo: consider use suitable error
-                |e| crate::error::Error::WriteError(format!("Fail to get result: {e}")),
-            )?;
-            results.extend(offsets?);
-        }
-        Ok(results)
-    }
-
-    fn prepare_list_offsets_requests(
-        &self,
-        table_id: i64,
-        partition_id: Option<i64>,
-        buckets: Vec<i32>,
-        offset_spec: OffsetSpec,
-    ) -> Result<HashMap<i32, ListOffsetsRequest>> {
-        let cluster = self.metadata.get_cluster();
-        let mut node_for_bucket_list: HashMap<i32, Vec<i32>> = HashMap::new();
-
-        for bucket_id in buckets {
-            let table_bucket = TableBucket::new(table_id, bucket_id);
-            let leader = cluster.leader_for(&table_bucket).ok_or_else(|| {
-                // todo: consider use another suitable error
-                crate::error::Error::InvalidTableError(format!(
-                    "No leader found for table bucket: table_id={table_id}, bucket_id={bucket_id}"
-                ))
-            })?;
-
-            node_for_bucket_list
-                .entry(leader.id())
-                .or_default()
-                .push(bucket_id);
-        }
-
-        let mut list_offsets_requests = HashMap::new();
-        for (leader_id, bucket_ids) in node_for_bucket_list {
-            let request =
-                ListOffsetsRequest::new(table_id, partition_id, bucket_ids, offset_spec.clone());
-            list_offsets_requests.insert(leader_id, request);
-        }
-        Ok(list_offsets_requests)
-    }
-
-    async fn send_list_offsets_request(
-        &self,
-        request_map: HashMap<i32, ListOffsetsRequest>,
-    ) -> Result<Vec<JoinHandle<Result<HashMap<i32, i64>>>>> {
-        let mut tasks = Vec::new();
-
-        for (leader_id, request) in request_map {
-            let rpc_client = self.conns.clone();
-            let metadata = self.metadata.clone();
-
-            let task = tokio::spawn(async move {
-                let cluster = metadata.get_cluster();
-                let tablet_server = cluster.get_tablet_server(leader_id).ok_or_else(|| {
-                    // todo: consider use more suitable error
-                    crate::error::Error::InvalidTableError(format!(
-                        "Tablet server {leader_id} not found"
-                    ))
-                })?;
-                let connection = rpc_client.get_connection(tablet_server).await?;
-                let list_offsets_response = connection.request(request).await?;
-                list_offsets_response.offsets()
-            });
-            tasks.push(task);
-        }
-
-        Ok(tasks)
     }
 
     async fn poll_for_fetches(&self) -> Result<HashMap<TableBucket, Vec<ScanRecord>>> {
